Split ir-dae function-owner tests; flatten test-side lint debt
Two verified-zero-semantic-change families in the ir-dae test tree,
combined because they touch the same files: (1) the SPEC_0021 split —
function_owners.rs (2146 lines) moves its tensor-builtin block verbatim
into the new child module function_tensor_builtins.rs (machine-verified
zero non-blank line loss, both files under the limit); (2) the
denied-lint repairs in function_conditionals.rs, function_owners.rs,
model_event_transactions.rs, and tests.rs (fixture-builder and
assertion-block extraction, one flat_map rewrite preserving row-major
order — no assertion weakened, no test removed). Containing-tree
verification under the pinned Nix env: cargo test -p rumoca-ir-dae
green (362 passed across the ir crate pair), cargo clippy
--all-targets --all-features -D warnings exit 0, fmt clean.
