+++ crates/rumoca-ir-dae/src/tests/function_conditionals.rs
@@ -34,20 +34,7 @@ fn a_conditional_that_joins_no_value_is_admitted_and_stores_no_assignment_group(
                 dae.functions(|functions| {
                     functions.assign(&mut body, output, zero, assignment_at)
                 })?;
-                let domain = dae.domains(|domains| {
-                    domains.structured(
-                        StructuredIndexDomain {
-                            binders: vec![StructuredIndexBinder {
-                                id: 0,
-                                display_name: "k".to_string(),
-                                lower: 1,
-                                upper: 0,
-                                step: 1,
-                            }],
-                        },
-                        loop_at,
-                    )
-                })?;
+                let domain = empty_loop_domain(dae, loop_at)?;
                 let mut loop_body = dae
                     .functions(|functions| functions.begin_loop(body, domain, [output], loop_at))?;
                 let k = dae.expressions(|expressions| {
@@ -65,33 +52,7 @@ fn a_conditional_that_joins_no_value_is_admitted_and_stores_no_assignment_group(
                         .at(condition_at)
                         .binary(BinaryOperator::Greater, k, bound)
                 })?;
-                // A forged join keeps its rectangularity proof: an empty target
-                // list admits only empty branches and an empty fallback.
-                for (branches, fallback) in
-                    [(vec![vec![k]], Vec::new()), (vec![Vec::new()], vec![k])]
-                {
-                    let forged = dae.functions(|functions| {
-                        functions.assign_conditional_all_loop(
-                            &mut loop_body,
-                            &[],
-                            &[condition],
-                            &branches,
-                            &fallback,
-                            conditional_at,
-                        )
-                    });
-                    assert!(
-                        matches!(
-                            forged,
-                            Err(DaeConstructionError::InvalidArity {
-                                expected: 0,
-                                found: 1,
-                                ..
-                            })
-                        ),
-                        "a forged vacuous join must keep its arity proof: {forged:?}"
-                    );
-                }
+                assert_forged_vacuous_joins(dae, &mut loop_body, k, condition, conditional_at);
                 let no_condition = dae.functions(|functions| {
                     functions.assign_conditional_all_loop(
                         &mut loop_body,
@@ -129,39 +90,95 @@ fn a_conditional_that_joins_no_value_is_admitted_and_stores_no_assignment_group(
     })
     .expect("a conditional over a statically empty domain constructs vacuously");
 
-    let inspect = |view: DaeView<'_>| {
-        let function = view.function(view.function_id(0).unwrap()).unwrap();
-        let statements = function.statements().collect::<Vec<_>>();
-        let FunctionStatementView::For {
-            fold, statements, ..
-        } = statements[1].clone()
-        else {
-            panic!("the empty-domain loop stays a compact fold");
-        };
-        assert_eq!(
-            statements.count(),
-            0,
-            "a vacuous conditional owns no assignment group"
-        );
-        let fold = view.function_fold(fold).unwrap();
-        assert_eq!(view.domain(fold.domain()).unwrap().scalar_count(), 0);
-        // The carried value leaves the fold exactly as it entered it.
-        assert_eq!(fold.update_values().len(), 1);
-        assert_eq!(
-            fold.update_values().rhs(0),
-            fold.parameter_values().rhs(0),
-            "an empty domain carries its value through unchanged"
-        );
-    };
-    dae.inspect(inspect);
+    dae.inspect(assert_vacuous_conditional);
     let encoded = serde_json::to_string(&dae).unwrap();
     serde_json::from_str::<Dae>(&encoded)
         .unwrap()
-        .inspect(inspect);
+        .inspect(assert_vacuous_conditional);
     let binary = bincode::serialize(&dae).unwrap();
     bincode::deserialize::<Dae>(&binary)
         .unwrap()
-        .inspect(inspect);
+        .inspect(assert_vacuous_conditional);
+}
+
+/// The statically empty `for k in 1:0` domain the vacuous conditional lives in.
+fn empty_loop_domain<'dae>(
+    dae: &mut DaeConstruction<'dae>,
+    loop_at: DaeProvenance,
+) -> Result<DomainId<'dae>, DaeConstructionError> {
+    dae.domains(|domains| {
+        domains.structured(
+            StructuredIndexDomain {
+                binders: vec![StructuredIndexBinder {
+                    id: 0,
+                    display_name: "k".to_string(),
+                    lower: 1,
+                    upper: 0,
+                    step: 1,
+                }],
+            },
+            loop_at,
+        )
+    })
+}
+
+fn assert_forged_vacuous_joins<'dae>(
+    dae: &mut DaeConstruction<'dae>,
+    loop_body: &mut FunctionLoop<'dae>,
+    k: ExprId<'dae>,
+    condition: ExprId<'dae>,
+    conditional_at: DaeProvenance,
+) {
+    // A forged join keeps its rectangularity proof: an empty target
+    // list admits only empty branches and an empty fallback.
+    for (branches, fallback) in [(vec![vec![k]], Vec::new()), (vec![Vec::new()], vec![k])] {
+        let forged = dae.functions(|functions| {
+            functions.assign_conditional_all_loop(
+                loop_body,
+                &[],
+                &[condition],
+                &branches,
+                &fallback,
+                conditional_at,
+            )
+        });
+        assert!(
+            matches!(
+                forged,
+                Err(DaeConstructionError::InvalidArity {
+                    expected: 0,
+                    found: 1,
+                    ..
+                })
+            ),
+            "a forged vacuous join must keep its arity proof: {forged:?}"
+        );
+    }
+}
+
+fn assert_vacuous_conditional(view: DaeView<'_>) {
+    let function = view.function(view.function_id(0).unwrap()).unwrap();
+    let statements = function.statements().collect::<Vec<_>>();
+    let FunctionStatementView::For {
+        fold, statements, ..
+    } = statements[1].clone()
+    else {
+        panic!("the empty-domain loop stays a compact fold");
+    };
+    assert_eq!(
+        statements.count(),
+        0,
+        "a vacuous conditional owns no assignment group"
+    );
+    let fold = view.function_fold(fold).unwrap();
+    assert_eq!(view.domain(fold.domain()).unwrap().scalar_count(), 0);
+    // The carried value leaves the fold exactly as it entered it.
+    assert_eq!(fold.update_values().len(), 1);
+    assert_eq!(
+        fold.update_values().rhs(0),
+        fold.parameter_values().rhs(0),
+        "an empty domain carries its value through unchanged"
+    );
 }
 
 #[test]
